testando

+++ src/questionario/questionario.service.ts
@@ -6,7 +6,7 @@ import { PrismaClient } from '@prisma/client';
 @Injectable()
 export class QuestionarioService {
   constructor(private readonly prisma: PrismaClient) { }
-  async create(createQuestionarioDto: CreateQuestionarioDto, equipeMedicaId: string) {
+  async create(createQuestionarioDto: CreateQuestionarioDto, equipeMedica: string) {
     try {
       await this.prisma.questionario.create({
         data: {
@@ -21,7 +21,7 @@ export class QuestionarioService {
           },
           equipeMedica: {
             connect: {
-              id: equipeMedicaId
+              id: equipeMedica
             }
           }
         }
@@ -35,8 +35,13 @@ export class QuestionarioService {
     return undefined;
   }
 
-  async findAll() {
+  async findAll(equipeMedica: string) {
     return await this.prisma.questionario.findMany({
+      where: {
+        equipeMedica: {
+          id: equipeMedica
+        }
+      },
       select: {
         criadoEm: true,
         pergunta: true,
@@ -51,24 +56,71 @@ export class QuestionarioService {
       }
     });
   }
+  
+  // async update(id: string, updateQuestionarioDto: UpdateQuestionarioDto) {
+  //   try {
 
-  findOne(id: number) {
-    return `This action returns a #${id} questionario`;
-  }
+  //     if (!updateQuestionarioDto) {
+  //       throw new HttpException("É necessário informar uma pergunta ou alternativas", HttpStatus.BAD_REQUEST);
+  //     }
 
-  update(id: number, updateQuestionarioDto: UpdateQuestionarioDto) {
-    return `This action updates a #${id} questionario`;
-  }
+  //     const questionarioExiste = await this.prisma.questionario.findUnique({
+  //       where: { id },
+  //       include: { alternativas: true }
+  //     });
 
-  async removeQuestionario(questionarioId: string) {
+  //     if (!questionarioExiste) {
+  //       throw new HttpException("Questionário não encontrado", HttpStatus.NOT_FOUND);
+  //     }
+
+  //     await this.prisma.alternativa.deleteMany({
+  //       where: { questionarioId: id }
+  //     });
+
+  //     return await this.prisma.questionario.update({
+  //       where: { id },
+  //       data: {
+  //         pergunta: updateQuestionarioDto.pergunta,
+  //         peso: updateQuestionarioDto.peso,
+  //         observacao: updateQuestionarioDto.observacao,
+  //         alternativas: updateQuestionarioDto.alternativas ? {
+  //           update: updateQuestionarioDto.alternativas.map((alternativa) => ({
+  //             where: { id: alternativa.id },
+  //             data: {
+  //               alternativa: alternativa.alternativa,
+  //               peso: alternativa.peso,
+  //             }
+  //           }))
+  //         } : undefined
+  //       },
+  //       include: {
+  //         alternativas: true
+  //       }
+  //     });
+  //   } catch (error) {
+  //     if (error instanceof HttpException) {
+  //       throw error;
+  //     }
+  //     console.log(error)
+  //     throw new HttpException("Erro interno do servidor", HttpStatus.INTERNAL_SERVER_ERROR);
+  //   }
+  // }
+
+  async removeQuestionario(questionarioId: string, equipeMedica: string) {
     const questionarioExiste = await this.prisma.questionario.findUnique({
       where: {
         id: questionarioId
       }
     })
+
     if (!questionarioExiste) {
       throw new HttpException("Questionário não encontrado", HttpStatus.NOT_FOUND);
     }
+
+    if (questionarioExiste.equipeMedicaId !== equipeMedica) {
+      throw new HttpException("Você não tem permissão para deletar este questionário", HttpStatus.FORBIDDEN);
+    }
+
     await this.prisma.questionario.delete({
       where: {
         id: questionarioId
